unlock a profile
setState() callback in Tenants reaches the tab tenant passing from all the widgets in the middle. When the button (unlock the profile) is pressed the set state is invoked and the value "match" is spread in all the classes that use it. Each classes manage the UI checking the value of "match"

After having unlocked a profile, it's no more possible to swipe in the ignore direction

fixed some errors